Pull everything up into LatestDagBase
